reafctor streamMessageCallback util to hook
