+++ template/src/components/RTMConfigure.tsx
@@ -84,8 +84,8 @@ const RtmConfigure = (props: any) => {
 
   React.useEffect(() => {
     const handBrowserClose = (ev) => {
-      ev.preventDefault();
-      return (ev.returnValue = 'Are you sure you want to exit?');
+      //  ev.preventDefault();
+      //  return (ev.returnValue = 'Are you sure you want to exit?');
     };
     const logoutRtm = () => {
       engine.current.leaveChannel(rtcProps.channel);
