+++ template/src/subComponents/caption/Caption.tsx
@@ -1,60 +1,48 @@
 import {StyleSheet, View, Platform} from 'react-native';
 import React, {MutableRefObject} from 'react';
-import {isWeb, useRtc} from 'customization-api';
+import {isWeb, useRender, useRtc} from 'customization-api';
 import protoRoot from './proto/ptoto';
 import {useCaption} from './useCaption';
 import {TranscriptText} from './TranscriptText';
 import Spacer from '../../../src/atoms/Spacer';
 import Loading from '../Loading';
-import {RenderObjects} from '../../../agora-rn-uikit/src/Contexts/RtcContext';
-import {streamMessageCallback} from './utils';
+// import {streamMessageCallback} from './utils';
 import {isWebInternal} from '../../utils/common';
+import useStreamMessageUtils from './useStreamMessageUtils';
+import {StreamMessageCallback} from 'react-native-agora/lib/typescript/common/RtcEvents';
 
-interface CaptionProps {
-  renderListRef: MutableRefObject<{renderList: RenderObjects}>;
-}
-
-const Caption: React.FC<CaptionProps> = ({renderListRef}) => {
+const Caption: React.FC = () => {
   const {RtcEngine} = useRtc();
-  const finalList = React.useRef<{[key: number]: string[]}>({}); // holds transcript of final words of all users
   const {
-    setMeetingTranscript,
-    meetingTranscript,
     isLangChangeInProgress,
-    isSTTActive,
     captionObj, //state for current live caption for all users
     setCaptionObj,
     isSTTListenerAdded,
     setIsSTTListenerAdded,
   } = useCaption();
-  const startTimeRef = React.useRef<number>(0);
-  const meetingTextRef = React.useRef<string>(''); // This is the full meeting text concatenated together.
 
-  const meetingTranscriptRef = React.useRef(meetingTranscript);
-  const sttObj = {
-    renderListRef,
-    finalList,
-    meetingTextRef,
-    startTimeRef,
-    meetingTranscriptRef,
-    setMeetingTranscript,
-    setCaptionObj,
-  };
+  const {streamMessageCallback} = useStreamMessageUtils();
+  const {renderList} = useRender();
 
-  const handleStreamMessageCallback1 = (...args: any[]) => {
+  const handleStreamMessageCallback1 = (
+    ...args: [number, Uint8Array] | [number, string, Uint8Array]
+  ) => {
     setIsSTTListenerAdded(true);
     if (isWebInternal()) {
-      streamMessageCallback(args, sttObj);
+      streamMessageCallback(args as [number, Uint8Array]);
     } else {
       const [uid, , data] = args;
       const streamBuffer = Object.values(data);
-      streamMessageCallback([uid, streamBuffer], sttObj);
+      streamMessageCallback([uid, new Uint8Array(streamBuffer)]);
     }
   };
 
   React.useEffect(() => {
     if (!isSTTListenerAdded) {
-      RtcEngine.addListener('StreamMessage', handleStreamMessageCallback1);
+      RtcEngine.addListener(
+        'StreamMessage',
+        handleStreamMessageCallback1 as unknown as StreamMessageCallback,
+      );
     }
     setCaptionObj({}); // clear live captions on mount
   }, []);
@@ -69,9 +57,7 @@ const Caption: React.FC<CaptionProps> = ({renderListRef}) => {
       {speakers.map(([key, value], index) => (
         <TranscriptText
           key={key}
-          user={
-            renderListRef.current.renderList[Number(key)]?.name || 'Speaker'
-          }
+          user={renderList[Number(key)]?.name || 'Speaker'}
           value={value.trim()}
           captionContainerStyle={
             activeSpeakers.length === 1
