+++ template/src/subComponents/caption/CaptionContainer.tsx
@@ -8,7 +8,7 @@ import {calculatePosition, isMobileUA, useIsDesktop} from '../../utils/common';
 import IconButton from '../../../src/atoms/IconButton';
 import hexadecimalTransparency from '../../../src/utils/hexadecimalTransparency';
 import ActionMenu, {ActionMenuItem} from '../../../src/atoms/ActionMenu';
-import {SidePanelType, useRender, useSidePanel} from 'customization-api';
+import {useSidePanel} from 'customization-api';
 import LanguageSelectorPopup, {getLanguageLabel} from './LanguageSelectorPopup';
 import useSTTAPI from './useSTTAPI';
 import events, {EventPersistLevel} from '../../rtm-events-api';
@@ -19,18 +19,12 @@ import {useMeetingInfo} from 'customization-api';
 const CaptionContainer = () => {
   const {isCaptionON, setIsCaptionON} = useCaption();
   const moreIconRef = React.useRef<View>(null);
-  const {renderList} = useRender();
   const [actionMenuVisible, setActionMenuVisible] =
     React.useState<boolean>(false);
-  const renderListRef = React.useRef({renderList});
   const isDesktop = useIsDesktop();
   const {
     data: {isHost},
   } = useMeetingInfo();
-
-  React.useEffect(() => {
-    renderListRef.current.renderList = renderList;
-  }, [renderList]);
 
   return isCaptionON ? (
     <View
@@ -49,7 +43,7 @@ const CaptionContainer = () => {
           setActionMenuVisible={setActionMenuVisible}
         />
       )}
-      <Caption renderListRef={renderListRef} />
+      <Caption />
     </View>
   ) : null;
 };
